Add files via upload
The enemy sights will be changed. This one has the first three sights that extend from the enemy, then that changes to the 8 directional sights that never seemed to work porperly. After this I will change the sights to one vertical line that scans up and down in a semi-circle